+++ config/routes.rb
@@ -11,9 +11,8 @@ Rails.application.routes.draw do
     get :research
     get :services
     get :sitemap_xml, path: "sitemap.xml.gz"
-    get :submit_contact, to: redirect("contact")
     get :thanks
     get :version
-    post :submit_contact
+    post :submit_contact, path: "contact"
   end
 end